Reduce prevalence of unwanted stereo effect
Change pluck method to feed the same random input into both channels.
This helps prevent a bug where there is a difference of output between
channels despite having the same input in both channels. Remove the
extra popSample in pluck method as it is unnecessary. Add LOOP macro
for generic looping that doesn't need access to incremented variable
as it may be useful elsewhere.